Update 6nd Gate for SPFI ---Non-Fish Module Update for Training (July 23, 2014)---x Purchasing Module Update
Updates:
+ Purchase Requisition
	/ Printout PRS Form to A4

+ Supplier Assignment
	/ Item order to assign as prs_code ascending

+ Purchase Order
	/ Takeout Currency Rate in Purchase Order
	/ Printout PO Form to Letter with Macro to adjust form

+++ src/java/com/app/wms/engine/db/dao/spring/PrsDetailDaoImpl.java
@@ -48,7 +48,7 @@ public class PrsDetailDaoImpl extends AbstractDAO implements ParameterizedRowMap
         su.declareParameter(new SqlParameter(java.sql.Types.VARCHAR));
         su.declareParameter(new SqlParameter(java.sql.Types.VARCHAR));
         su.declareParameter(new SqlParameter(java.sql.Types.VARCHAR));
-        su.declareParameter(new SqlParameter(java.sql.Types.INTEGER));
+        su.declareParameter(new SqlParameter(java.sql.Types.NUMERIC));
         su.declareParameter(new SqlParameter(java.sql.Types.VARCHAR));
         su.declareParameter(new SqlParameter(java.sql.Types.NUMERIC));
         su.compile();
